services/order/requirements.txt
fastapi
python-dotenv
sqlalchemy
pydantic
requests
uvicorn
pytest-bdd
python-dotenv
